Update event information to be split into event data and event metadata:
The Logger::event() API accepts a event data HashRef and a optional
metadata HashRef.

It then builds the unified internal event structure, a HashRef with two
keys: 'e' to keep the event data, and 'm' to keep the event metadata.

The Logan::dispatch() API still receives the event data as the first
parameter, but now it also has access to the internal event metadata as
the second parameter.

Signed-off-by: Pedro Melo <melo@simplicidade.org>